requirements-bk.txt
#adium-theme-ubuntu==0.3.4
attrs==15.2.0
Babel==2.3.4
bcrypt==3.1.3
beautifulsoup4==4.4.1
bzr==2.7.0
#BzrTools==2.6.0
certifi==2017.4.17
cffi==1.10.0
Chameleon==2.24
chardet==3.0.3
configobj==5.0.6
cryptography==1.2.3
damm==1.0
decorator==4.0.10
docutils==0.12
ebaysdk==2.1.4
ecdsa==0.13
enum34==1.1.2
feedparser==5.2.1
funcsigs==1.0.2
gdata==2.0.18
gevent==1.1.2
greenlet==0.4.10
#gyp==0.1
html5lib==0.999
httpagentparser==1.8.0
httplib2==0.9.1
idna==2.5
ipaddress==1.0.16
jcconv==0.2.3
Jinja2==2.8
keyring==7.3
launchpadlib==1.10.3
#lazr.restfulclient==0.13.4
lazr.uri==1.0.3
linecache2==1.0.0
#lptools==0.2.0
lxml==3.5.0
Mako==1.0.4
MarkupSafe==0.23
mock==2.0.0
ndg-httpsclient==0.4.0
oauth==1.0.1
ofxparse==0.15
openerp-web==7.0.406
openerp-web-kanban==7.0.406
#PAM==0.4.2
paramiko==2.2.1
Parsley==1.2
passlib==1.6.5
pbr==1.8.0
#pdftools==0.37
pexpect==4.3.0
Pillow==3.4.1
psutil==4.3.1
psycogreen==1.0
psycopg2==2.6.1
ptyprocess==0.5.2
pyasn1==0.1.9
pyasn1-modules==0.0.7
PyChart==1.39
pycparser==2.18
pycrypto==2.6.1
#pycurl==7.43.0
pydot==1.2.3
Pygments==2.1
#pygobject==3.20.0
pygpgme==0.3
pyinotify==0.9.6
pyliblzma==0.5.3
pymongo==3.6.0
PyNaCl==1.1.2
pyodbc==3.1.1
pyOpenSSL==0.15.1
pyparsing==2.1.10
pyPdf==1.13
PyPDF2==1.26.0
pyserial==3.1.1
pysftp==0.2.9
#python-apt==1.1.0b1
Python-Chart==1.39
python-dateutil==2.5.3
python-ldap==2.4.22
python-openid==2.2.5
python-stdnum==1.2
pytz==2016.7
pyusb==1.0.0
PyWebDAV==0.9.8
pyxdg==0.25
PyYAML==3.12
qrcode==5.3
reportlab==3.3.0
requests==2.11.1
roman==2.0.0
SecretStorage==2.1.3
service-identity==16.0.0
simplejson==3.5.3
six==1.10.0
stevedore==1.12.0
suds-jurko==0.6
#torbrowser-launcher==0.2.4
traceback2==1.4.0
Twisted==16.0.0
#txsocksx==0.0.0
unicodecsv==0.14.1
unittest2==0.5.1
#unity-lens-photos==1.0
urllib3==1.21.1
uTidylib==0.2
validator==2.0.6
vatnumber==1.2
virtualenv==15.0.1
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.1
vobject==0.9.3
wadllib==1.3.2
Wand==0.4.4
Werkzeug==0.9.4
wkhtmltopdf==0.2
#wkhtmltopdf-pack==0.12.3.0
XlsxWriter==0.9.3
xlwt==1.1.2
zope.interface==4.1.3
ZSI==2.1a1
